remove loader and fade out  (delayed) no results too
