+++ packages/dataviews/src/dataviews/style.scss
@@ -4,7 +4,6 @@
 
 .dataviews-wrapper,
 .dataviews-picker-wrapper {
-	position: relative;
 	height: 100%;
 	overflow: auto;
 	box-sizing: border-box;
@@ -16,17 +15,6 @@
 	font-size: $default-font-size;
 	line-height: $default-line-height;
 	background-color: var(--wp-dataviews-color-background, $white);
-}
-
-.dataviews-wrapper .dataviews-wrapper__loading-progress {
-	position: fixed;
-	top: 0;
-	left: 0;
-	right: 0;
-	width: 100%;
-	z-index: 1000;
-	border-radius: 0;
-	--wp-components-color-foreground: var(--wp-admin-theme-color);
 }
 
 .dataviews__view-actions,
@@ -54,6 +42,10 @@
 	@media not (prefers-reduced-motion) {
 		transition: padding ease-out 0.1s;
 	}
+}
+
+.dataviews-no-results.is-refreshing {
+	opacity: 0.5;
 }
 
 .dataviews-loading-more {
